moved HTTPNetworkingClientError to its own file + tests
